+++ src/compas/geometry/_core/angles.py
@@ -81,14 +81,14 @@ def angle_vectors_signed(u, v, normal, deg=False, threshold=1e-3):
     >>> normal = [0.0, 0.0, 1.0]
     >>> angle_vectors_signed([0.0, 1.0, 0.0], [1.0, 0.0, 0.0], normal)
     -1.57079
+
     """
     angle = angle_vectors(u, v)
     normal_uv = cross_vectors(u, v)
 
     if length_vector(normal_uv) > threshold:
-        # check if normal_uv has the same direction as normal
-        angle_btw_normals = angle_vectors(normal, normal_uv)
-        if angle_btw_normals > threshold:
+        # check if normal_uv has the same direction as normals
+        if dot_vectors(normal, normal_uv) < -threshold:
             angle *= -1
 
     if deg: